+++ Commands/Base/GetAuthenticationRealm.cs
@@ -8,8 +8,9 @@ namespace OfficeDevPnP.PowerShell.Commands
     [Cmdlet(VerbsCommon.Get, "SPOAuthenticationRealm")]
     [CmdletHelp("Gets the authentication realm for the current web", 
         Category = CmdletHelpCategory.Base)]
-    [CmdletExample(Code = @"
-PS:> Get-SPOAuthenticationRealm -Url https://contoso.sharepoint.com", SortOrder = 1)]
+    [CmdletExample(
+        Code = @"PS:> Get-SPOAuthenticationRealm -Url https://contoso.sharepoint.com", 
+        SortOrder = 1)]
     public class GetAuthenticationRealm : SPOCmdlet
     {