+++ Commands/Properties/AssemblyInfo.cs
@@ -39,5 +39,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("2.0.1601.0")]
-[assembly: AssemblyFileVersion("2.0.1601.0")]
+[assembly: AssemblyVersion("2.1.1602.0")]
+[assembly: AssemblyFileVersion("2.1.1602.0")]